Implement disk usage and battery status
After discussing on GitHub, I have identified that the information we
require is the disk usage rather than memory usage. Additionally, we
need to retrieve battery status, including the remaining battery time
and whether the battery is currently charging or not. This will allow us
to determine whether or not to alert the user to charge their device.

I attempted to test the implementation, but I am currently in the
process of learning how to use 'mockall'.